compiles

+++ recursion/compiler/src/asm/instruction.rs
@@ -5,6 +5,7 @@ use core::fmt;
 use p3_field::{ExtensionField, PrimeField32};
 use sp1_recursion_core::cpu::Instruction;
 use sp1_recursion_core::runtime::Opcode;
+use sp1_recursion_core::runtime::PERMUTATION_WIDTH;
 
 use super::A0;
 use crate::util::canonical_i32_to_field;
@@ -716,7 +717,7 @@ impl<F: PrimeField32, EF: ExtensionField<F>> AsmInstruction<F, EF> {
                 Opcode::Poseidon2Compress,
                 i32_f(dst),
                 i32_f_arr(src),
-                i32_f_arr(src + PERMUTATION_WIDTH),
+                i32_f_arr(src + PERMUTATION_WIDTH as i32),
                 F::zero(),
                 F::zero(),
                 false,
